change to older version

+++ ofbiz/apache-ofbiz/framework/base/src/test/groovy/org/apache/ofbiz/base/util/FileUtilTests.groovy
@@ -21,28 +21,25 @@ package org.apache.ofbiz.base.util
 import org.apache.commons.io.FileUtils
 import org.junit.Test
 import org.junit.Ignore
-class FileUtilTests {
-
+public class FileUtilTests {
     /**
      * Test FileUtil zipFileStream and unzipFileToFolder methods, using README.adoc
      */
     @Test
     @Ignore("Skipping due to file not found issue")
     void zipReadme() {
-        String zipFilePath = UtilProperties.getPropertyValue('general', 'http.upload.tmprepository', 'runtime/tmp')
+        String zipFilePath = UtilProperties.getPropertyValue("general", "http.upload.tmprepository", "runtime/tmp")
         String zipName = 'README.adoc.zip'
         String fileName = 'README.adoc'
         File originalReadme = new File(fileName)
 
         //validate zipStream from README.adoc is not null
-        ByteArrayInputStream zipStream = FileUtil.zipFileStream(originalReadme.newInputStream(), fileName)
+        def zipStream = FileUtil.zipFileStream(originalReadme.newInputStream(), fileName)
         assert zipStream
 
         //ensure no zip already exists
         File readmeZipped = new File(zipFilePath, zipName)
-        if (readmeZipped.exists()) {
-            readmeZipped.delete()
-        }
+        if (readmeZipped.exists()) readmeZipped.delete()
 
         //write it down into tmp folder
         OutputStream out = new FileOutputStream(readmeZipped)
@@ -56,14 +53,11 @@ class FileUtilTests {
 
         //ensure no README.adoc exist in tmp folder
         File readme = new File(zipFilePath, fileName)
-        if (readme.exists()) {
-            readme.delete()
-        }
+        if (readme.exists()) readme.delete()
 
         //validate unzip and compare the two files
         FileUtil.unzipFileToFolder(readmeZipped, zipFilePath, false)
 
         assert FileUtils.contentEquals(originalReadme, new File(zipFilePath, fileName))
     }
-
 }
